📦 backup: src/util/text.py에 json 파싱 함수 추가.
